Create Details of the  manu & order table
View the menu_items table and write a query to find the number of items on the menu

What are the least and most expensive items on the menu?

How many Italian dishes are on the menu? What are the least and most expensive Italian dishes on the menu?

How many dishes are in each category? What is the average dish price within each category?

View the order_details table. What is the date range of the table?

How many orders were made within this date range? How many items were ordered within this date range?

Which orders had the most number of items?

How many orders had more than 12 items?